Added the Expenses Tracker program in Python
To run the program, please follow these steps:
1. Ensure you have the latest Python installed on your system as possible.
2. Download both files and make sure that both of them are in the same folder.
3. Open the project in your preferred IDE (e.g. VSCode, which is my IDE preference).
4. Run the program ExpensesTracker.py.

At the start of the program, If you want to test the program as a first-time user, type 'Y', add an initial balance as starting point, and follow the prompts to add an expense. This will generate a transaction file similar to the one attached. If you want to test the program with existing transaction records, type 'N', and then enter the name of the attached text file (e.g. "transactions_2024-07-15_16-13-34.txt").